split chat and app contexts
