+++ src/layouts/AnalyticsLayout.tsx
@@ -4,15 +4,12 @@ import {
 import classNames from "classnames";
 import Header from "../components/Header";
 import SettingsModal from "../components/Settings/SettingsModal";
-import { useThemeContext } from "../contexts/ThemeContext";
 
 type AnalyticsLayoutProps = {
     children: ReactElement;
 };
 
 const AnalyticsLayout = ({ children } : AnalyticsLayoutProps) => {
-    const { chatBackground, theme } = useThemeContext();
-
     return (
       <div className="relative flex h-full w-full">
         <div className="relative flex-1">
@@ -31,15 +28,7 @@ const AnalyticsLayout = ({ children } : AnalyticsLayoutProps) => {
             </div>
           </main>
           <div
-            className={classNames("absolute inset-0 mt-14 z-[-1]", {
-              "bg-cover bg-center": !!chatBackground,
-            })}
-            style={{
-              backgroundImage: chatBackground
-                ? `url(${chatBackground})`
-                : undefined,
-              filter: theme === "dark" ? "invert(1)" : undefined,
-            }}
+            className={classNames("absolute inset-0 mt-14 z-[-1]")}
           />
         </div>
